fix(db/migrate): turn DropIndexReceiptsOnTokenIdAndQuestionId into conditional migration

+++ db/migrate/20240901201521_drop_index_receipts_on_token_id_and_question_id.rb
@@ -1,5 +1,5 @@
 class DropIndexReceiptsOnTokenIdAndQuestionId < ActiveRecord::Migration[7.0]
   def change
-    remove_index :receipts, name: "index_receipts_on_token_id_and_question_id"
+    remove_index :receipts, name: "index_receipts_on_token_id_and_question_id" if index_exists?(:receipts, %i[token_id question_id], name: 'index_receipts_on_token_id_and_question_id')
   end
 end
